Move colorbar control to separate module
